fixes for v5r4
git-svn-id: http://svn.cern.ch/guest/lhcb/Boole/trunk@21011 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: d9fd954fe6e907da70e7f5aa8db33a4a9ca43e34

+++ Boole/Muon/MuonAlgs/MuonAlgs/MuonPhysicalChannelResponse.h
@@ -61,10 +61,10 @@ Rndm::Numbers* p_electronicNoise;
 int m_readoutIndex ;
 double m_minOfTimeJitter ;
 double m_maxOfTimeJitter ;
-float m_meadDeadtime ;
-float m_rmsDeadtime ;
-float m_chamberEfficiency ;
-float m_timeAdjustmentImprecision ;
+double m_meadDeadtime ;
+double m_rmsDeadtime ;
+double m_chamberEfficiency ;
+double m_timeAdjustmentImprecision ;
 MuonReadoutCond*  p_detectorResponse ;
  };
